--- bench_v1_test.go
@@ -1,312 +1,0 @@
-package protocol
-
-import (
-	"io"
-	"testing"
-
-	"encoding/json"
-)
-
-func benchPayload() []byte {
-	size := 256
-	var p []byte
-	for i := 0; i < size; i++ {
-		p = append(p, 'i')
-	}
-	return []byte(`{"input":"` + string(p) + `"}`)
-}
-
-var preparedPayload = benchPayload()
-
-func marshalProtobuf() ([]byte, *Reply, error) {
-	pub := &Publication{
-		Data: preparedPayload,
-	}
-	pushBytes, err := EncodePublicationPush(TypeProtobuf, "test", pub)
-	if err != nil {
-		return nil, nil, err
-	}
-	r := &Reply{
-		Result: pushBytes,
-	}
-	encoder := NewProtobufReplyEncoder()
-	res, err := encoder.Encode(r)
-	if err != nil {
-		return nil, nil, err
-	}
-	return res, r, nil
-}
-
-func marshalJSON() ([]byte, *Reply, error) {
-	pub := &Publication{
-		Data: preparedPayload,
-	}
-	pushBytes, err := EncodePublicationPush(TypeJSON, "test", pub)
-	if err != nil {
-		return nil, nil, err
-	}
-	r := &Reply{
-		Result: pushBytes,
-	}
-	encoder := NewJSONReplyEncoder()
-	res, err := encoder.Encode(r)
-	if err != nil {
-		return nil, nil, err
-	}
-	return res, r, nil
-}
-
-func BenchmarkReplyProtobufMarshal(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		d, r, err := marshalProtobuf()
-		if err != nil {
-			b.Fatal(err)
-		}
-		benchData = d
-		benchReply = r
-	}
-	b.ReportAllocs()
-}
-
-func BenchmarkReplyProtobufMarshalParallel(b *testing.B) {
-	b.ReportAllocs()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			d, r, err := marshalProtobuf()
-			if err != nil {
-				b.Fatal(err)
-			}
-			benchData = d
-			benchReply = r
-		}
-	})
-}
-
-func BenchmarkReplyProtobufUnmarshal(b *testing.B) {
-	params := &ConnectRequest{
-		Token: "token",
-	}
-	data, _ := params.MarshalVT()
-	cmd := &Command{
-		Id:     1,
-		Method: Command_CONNECT,
-		Params: data,
-	}
-	encoder := NewProtobufCommandEncoder()
-	data, _ = encoder.Encode(cmd)
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		benchConnectRequest = unmarshalProtobuf(b, data)
-	}
-	b.ReportAllocs()
-}
-
-func BenchmarkReplyProtobufUnmarshalParallel(b *testing.B) {
-	params := &ConnectRequest{
-		Token: "token",
-	}
-	data, _ := params.MarshalVT()
-	cmd := &Command{
-		Id:     1,
-		Method: Command_CONNECT,
-		Params: data,
-	}
-	encoder := NewProtobufCommandEncoder()
-	data, _ = encoder.Encode(cmd)
-	b.ResetTimer()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			benchConnectRequest = unmarshalProtobuf(b, data)
-		}
-	})
-	b.ReportAllocs()
-}
-
-func unmarshalProtobuf(b *testing.B, data []byte) *ConnectRequest {
-	decoder := GetCommandDecoder(TypeProtobuf, data)
-	defer PutCommandDecoder(TypeProtobuf, decoder)
-	cmd, err := decoder.Decode()
-	if err != nil && err != io.EOF {
-		b.Fatal(err)
-	}
-	if cmd == nil {
-		b.Fatal("nil command")
-	}
-	paramsDecoder := NewProtobufParamsDecoder()
-	req, err := paramsDecoder.DecodeConnect(cmd.Params)
-	if err != nil {
-		b.Fatal()
-	}
-	if req.Token != "token" {
-		b.Fatal()
-	}
-	return req
-}
-
-func BenchmarkReplyJSONMarshal(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		d, r, err := marshalJSON()
-		if err != nil {
-			b.Fatal(err)
-		}
-		benchData = d
-		benchReply = r
-	}
-	b.ReportAllocs()
-}
-
-func BenchmarkReplyJSONMarshalParallel(b *testing.B) {
-	b.ReportAllocs()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			d, r, err := marshalJSON()
-			if err != nil {
-				b.Fatal(err)
-			}
-			benchData = d
-			benchReply = r
-		}
-	})
-}
-
-func BenchmarkReplyJSONUnmarshal(b *testing.B) {
-	params := &ConnectRequest{
-		Token: "token",
-	}
-	data, _ := json.Marshal(params)
-	cmd := &Command{
-		Id:     1,
-		Method: Command_CONNECT,
-		Params: data,
-	}
-	encoder := NewJSONCommandEncoder()
-	data, err := encoder.Encode(cmd)
-	if err != nil {
-		b.Fatal(err)
-	}
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		benchConnectRequest = unmarshalJSON(b, data)
-	}
-	b.ReportAllocs()
-}
-
-func BenchmarkReplyJSONUnmarshalParallel(b *testing.B) {
-	params := &ConnectRequest{
-		Token: "token",
-	}
-	data, _ := json.Marshal(params)
-	cmd := &Command{
-		Id:     1,
-		Method: Command_CONNECT,
-		Params: data,
-	}
-	encoder := NewJSONCommandEncoder()
-	data, err := encoder.Encode(cmd)
-	if err != nil {
-		b.Fatal(err)
-	}
-	b.ResetTimer()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			benchConnectRequest = unmarshalJSON(b, data)
-		}
-	})
-	b.ReportAllocs()
-}
-
-func BenchmarkReplyJSONUnmarshalMultiple(b *testing.B) {
-	params := &ConnectRequest{
-		Token: "token",
-	}
-	data, _ := json.Marshal(params)
-	cmd := &Command{
-		Id:     1,
-		Method: Command_CONNECT,
-		Params: data,
-	}
-	encoder := NewJSONCommandEncoder()
-	data, err := encoder.Encode(cmd)
-	if err != nil {
-		b.Fatal(err)
-	}
-	data = append(data, []byte("\n")...)
-	data = append(data, data...)
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		unmarshalJSONMultiple(b, data)
-	}
-	b.ReportAllocs()
-}
-
-func BenchmarkReplyJSONUnmarshalMultipleParallel(b *testing.B) {
-	params := &ConnectRequest{
-		Token: "token",
-	}
-	data, _ := json.Marshal(params)
-	cmd := &Command{
-		Id:     1,
-		Method: Command_CONNECT,
-		Params: data,
-	}
-	encoder := NewJSONCommandEncoder()
-	data, err := encoder.Encode(cmd)
-	if err != nil {
-		b.Fatal(err)
-	}
-	data = append(data, []byte("\n")...)
-	data = append(data, data...)
-	b.ResetTimer()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			unmarshalJSONMultiple(b, data)
-		}
-	})
-	b.ReportAllocs()
-}
-
-func unmarshalJSON(b *testing.B, data []byte) *ConnectRequest {
-	decoder := GetCommandDecoder(TypeJSON, data)
-	defer PutCommandDecoder(TypeJSON, decoder)
-	cmd, err := decoder.Decode()
-	if (err != nil && err != io.EOF) || cmd == nil {
-		b.Fatal(err)
-	}
-	paramsDecoder := NewJSONParamsDecoder()
-	req, err := paramsDecoder.DecodeConnect(cmd.Params)
-	if err != nil {
-		b.Fatal()
-	}
-	if req.Token != "token" {
-		b.Fatal()
-	}
-	return req
-}
-
-func unmarshalJSONMultiple(b *testing.B, data []byte) {
-	decoder := GetCommandDecoder(TypeJSON, data)
-	defer PutCommandDecoder(TypeJSON, decoder)
-	cmd, err := decoder.Decode()
-	if err != nil {
-		b.Fatal(err)
-	}
-	paramsDecoder := NewJSONParamsDecoder()
-	req, err := paramsDecoder.DecodeConnect(cmd.Params)
-	if err != nil {
-		b.Fatal()
-	}
-	if req.Token != "token" {
-		b.Fatal()
-	}
-	cmd, err = decoder.Decode()
-	if (err != nil && err != io.EOF) || cmd == nil {
-		b.Fatal(err)
-	}
-	req, err = paramsDecoder.DecodeConnect(cmd.Params)
-	if err != nil {
-		b.Fatal()
-	}
-	if req.Token != "token" {
-		b.Fatal()
-	}
-}
