+++ client.proto
@@ -17,31 +17,12 @@ message EmulationRequest {
 }
 
 // Command sent from a client to a server.
-// ProtocolVersion1 uses id, method and params fields.
 // ProtocolVersion2 uses id and one of the possible request messages.
 message Command {
   // Id of command to let client match replies to commands.
   uint32 id = 1;
 
-  enum MethodType {
-    CONNECT = 0;
-    SUBSCRIBE = 1;
-    UNSUBSCRIBE = 2;
-    PUBLISH = 3;
-    PRESENCE = 4;
-    PRESENCE_STATS = 5;
-    HISTORY = 6;
-    PING = 7;
-    SEND = 8;
-    RPC = 9;
-    REFRESH = 10;
-    SUB_REFRESH = 11;
-  }
-  // Method is used in ProtocolVersion1 only.
-  MethodType method = 2;
-
-  // Params is used in ProtocolVersion1 only.
-  bytes params = 3;
+  reserved 2, 3;
 
   // ProtocolVersion2 client can send one of the following requests. Server will
   // only take the first non-null request out of these and may return an error if
@@ -62,7 +43,6 @@ message Command {
 }
 
 // Reply sent from a server to a client.
-// ProtocolVersion1 uses id, error and result fields.
 // ProtocolVersion2 uses id and one of the possible concrete result messages.
 message Reply {
   // Id will only be set to a value > 0 for replies to commands. For pushes
@@ -71,8 +51,7 @@ message Reply {
   // Error can only be set in replies to commands. For pushes it will have zero value.
   Error error = 2;
 
-  // Result is used in ProtocolVersion1 only.
-  bytes result = 3;
+  reserved 3;
 
   // ProtocolVersion2 server can send one of the following fields. We are not using
   // oneof here due to JSON interoperability concerns.
@@ -95,23 +74,8 @@ message Reply {
 // ProtocolVersion1 uses type, channel and data fields.
 // ProtocolVersion2 uses channel and one of the possible concrete push messages.
 message Push {
-  enum PushType {
-    PUBLICATION = 0;
-    JOIN = 1;
-    LEAVE = 2;
-    UNSUBSCRIBE = 3;
-    MESSAGE = 4;
-    SUBSCRIBE = 5;
-    CONNECT = 6;
-    DISCONNECT = 7;
-    REFRESH = 8;
-  }
-  // Type is used in ProtocolVersion1 only.
-  PushType type = 1;
+  reserved 1, 3;
   string channel = 2;
-
-  // Data is used in ProtocolVersion1 only.
-  bytes data = 3;
 
   // ProtocolVersion2 server can push one of the following fields to the client. We are
   // not using oneof here due to JSON interoperability concerns.
@@ -134,7 +98,7 @@ message ClientInfo {
 }
 
 message Publication {
-  // 1-3 skipped here for backwards compatibility.
+  reserved 1, 2, 3;
   bytes data = 4;
   ClientInfo info = 5;
   uint64 offset = 6;
@@ -150,14 +114,14 @@ message Leave {
 }
 
 message Unsubscribe {
-  // Field 1 removed (bool resubscribe).
+  reserved 1;
   uint32 code = 2;
   string reason = 3;
 }
 
 message Subscribe {
   bool recoverable = 1;
-  // 2-3 skipped here for backwards compatibility.
+  reserved 2, 3;
   string epoch = 4;
   uint64 offset = 5;
   bool positioned = 6;
@@ -228,7 +192,7 @@ message SubscribeRequest {
   string channel = 1;
   string token = 2;
   bool recover = 3;
-  // 4-5 skipped here for backwards compatibility.
+  reserved 4, 5;
   string epoch = 6;
   uint64 offset = 7;
   bytes data = 8;
@@ -241,7 +205,7 @@ message SubscribeResult {
   bool expires = 1;
   uint32 ttl = 2;
   bool recoverable = 3;
-  // 4-5 skipped here for backwards compatibility.
+  reserved 4, 5;
   string epoch = 6;
   repeated Publication publications = 7;
   bool recovered = 8;
@@ -298,7 +262,7 @@ message StreamPosition {
 
 message HistoryRequest {
   string channel = 1;
-  // 2-6 skipped here for backwards compatibility.
+  reserved 2, 3, 4, 5, 6;
   int32 limit = 7;
   StreamPosition since = 8;
   bool reverse = 9;
